pyproject.toml
[project]
name = "gemini-multimodal-api-calls"
version = "0.1.0"
description = "A demo repo to show how to call Gemini 2.0 Flash Experimental with media files"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.59.6",
    "python-dotenv>=1.0.1",
]
